// src/app/features/shared/models/error-codes/error-codes.enum.ts
export enum ErrorCode {
  GROUP_ADMIN_PRIVILEGES_REQUIRED = 1007,

  RESOURCE_OWNER_PRIVILEGES_REQUIRED = 1008,
  USER_HAS_NOT_ENOUGH_MONEY = 1009,

  VALIDATION_ERROR = 1501,
  INTERNAL_SERVER_ERROR = 1502,
  MALFORMED_REQUEST = 1503,
  MISSING_PARAMETER = 1504,
  METHOD_NOT_ALLOWED = 1505,
  UNSUPPORTED_MEDIA_TYPE = 1506,
  ACCESS_DENIED = 1507,
  TYPE_MISMATCH = 1508,
  MISSING_REQUEST_BODY = 1509,
  MISSING_REQUEST_COOKIE = 1510,
  MISSING_ACCESS_TOKEN_COOKIE = 1511,
  MISSING_REFRESH_TOKEN_COOKIE = 1512,
  OPTIMISTIC_LOCKING_FAILURE = 1513,

  INVALID_CREDENTIALS = 2001,
  TOKEN_EXPIRED = 2002,
  INVALID_TOKEN = 2003,
  REFRESH_TOKEN_EXPIRED = 2013,
  ACCESS_TOKEN_EXPIRED = 1019,
  ACCOUNT_LOCKED = 1020,
}